+++ site/astro.config.mjs
@@ -1,11 +1,16 @@
 import { defineConfig } from "astro/config";
 import react from "@astrojs/react";
+import sitemap from "@astrojs/sitemap";
 import tailwind from "@astrojs/tailwind";
 
 export default defineConfig({
   site: "https://ruslanbes.github.io",
   base: "/eurovision-top-20/",
-  integrations: [react(), tailwind({ applyBaseStyles: false })],
+  integrations: [
+    react(),
+    tailwind({ applyBaseStyles: false }),
+    sitemap(),
+  ],
   server: {
     port: 3420,
   },